moved save_load_and_test method in conftest
